Extract server logic from index.js, add minimal getopts support
